Humanoid animator manager
Fixed humanoid so max speed has a hardset limit of 1.6x greater than the set max speed in the event of running down very steep inclines. This is to keep the Animator smooth.

The new HumanoidAnim also assumes that the user has FinalIK installed. This can easily be removed though via the deletion of the component reference.