Fixes attacking so character can only damage when attack animation playing.
Added a temp 3D model to facilitate temporary animations for the attack
system. Now the character can only do damage during an anime with a
"Damage" attack curve inside of it. This is all managed by a new
CharacterAnimManager class. Which triggers and tracks the current attack
playing and if damage can be dealt to an enemy at that time.