Fixed advanced jumping and wall climbing
Gave more control to how high/short a jump is by modifying the extra
down force to only apply if not jumping upwards. Thereby allowing for
more control of jump height. Which results in the player climbing up
Ground surfaces smoother and sliding off of them with the added gravity
applied while not trying to jump.