Character Attack Combo Damage & Damage Limit Move
The combo attack animation can now deal a stream of constant damage by
toggling the DoDamage parameter inside the animation's curve. Allowing
the damage dealt to match the animation itself. The penalty for damage dealt and
knockback can be adjusted in the Weapon's inspector settings. Each
"Weapon" can also have a maximum amount of combo damage. When reached
the combo is stopped and a heavy attack plays to finish off the combo.