Alternative std attack & support for soft combos
When tapping the standard attack button, its animation will alternate
between a left and right handed punch animation. This system enables
chains of animations to be used as combos. However, the old alternative
hit system needs to be removed as this new one replaces that ones
intended implementation. Furthermore, the ability to do a "charged"
attack after transitioning to an alt handed attack Must be added.